api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
aiofiles==24.1.0
python-docx==1.1.2
PyPDF2==3.0.1
nltk==3.9.1
textstat==0.7.4
python-frontmatter==1.1.0
httpx==0.28.1
pytest==8.3.4
